Reduce tolerances and mark estimate_emos_coefficients tests as xfail
nccmp expresses relative tolerances as percentages, which is different
to numpy.allclose combined relative and absolute tolerances.
Adjusting the tolerance constants to be the same as before causes two
estimate_emos_coefficients tests to fail and inspection of the numbers
output shows differences which may be scientifically significant.
These different outputs will need to be investigated and scientifically
appropriate tolerances determined for these tests.