Add separate test for checksums across all KGOs
This test uses its own code to handle KGO re-creation, which in this
case means regenerating the checksum file.
Re-generation of the checksum file being centralised in one test case
avoids race conditions on updating the checksum file as part of KGO
re-creation in many tests which could be run in parallel such as via
pytest-xdist.