IMPRO-1725: Provide a list of ignorable attributes to the netcdf compare utility (#1266)
* Provide a list of ignorable attributes to the netcdf compare utilites.

This change enables us to set a variety of attributes as ignorable, 
which will reduce the number of times we must update acceptance test 
data for trivial changes that are unimportant for checking data 
integrity.

* Interface changes etc. following review.

Reordered interface to place reporter last. Removed named passing of 
ignored_attributes argument. Updated the CLI calling compare to provide 
access to the ignored_attributes option. Added an additional acceptance 
test to cover this.

* Update interfaces for tests now that ignored_attributes is positional.

* Review changes.