Modify import and test skipping for timezonefinder module (#1306)
* Modify import and test skipping for timezonefinder module.

The timezonefinder module is imported in the init of the 
GenerateTimezoneMask plugin. This is instantiated for all the unit and 
acceptance test, so they must all be skipped if the module is not 
available. Pylint further checks that imports are available, so to work 
around this a check has been added to the import that satisfies pylint 
in the absence of the module.

* Move to use importlib instead of imp.