Added the forecast period bounds to emos coefficients cubes.
This is required as the entire coordinate is checked when applying the
coefficients. An alternative approach would be to only check the
coordinate point, but this approach has the benefit of ensuring, e.g.
that we cannot apply an the coefficients for an instantaneous
temperature to a period temperature minimum/maxmimum.