A small bug fix and some tidying.
I noted that I was using the neighbourhood value rather than the cell 
value to determine whether ratio should take the result of the division 
or not. This has been changed after checking in the notebook that it had 
negligible effect. I have extended the doc-strings. I have also moved 
the cube creation to the end of the calculate_ratio method as it was 
previously a bit unconventional.