Modifications to cube combiner to keep probabilistic naming elements.
These probabilistic naming elements are only preserved when broadcasting
to a threshold coordinate has been used. This ensures that the
information about how the data has been thresholded is preserved. The
naming elements are taken from the first cube in the input cubelist
which is always expected to include the coordinate to which broadcasting
is happening for subsequent cubes.