Update IMPROVER environment and tests (#1424)
* Fix proj vesion to 0.7.0 in py36 env.

When building the improver_py36_iris22 environment, the conda solver 
does not consistently retrieve a particular version of proj upon which 
cartopy relies. If proj 0.7.2 is chosen, the resulting cartopy build 
will not function. This PR fixes the proj version to one at which 
cartopy works.

* Remove reliance on counting number of warnings from spot-data acceptance tests.

* Regenerate checksums.

* Black.

* Improve test doc-string.

* Update all cache numbers to point at version 2.

* Modifying environment proj version somehow results in new sphinx version (3.5.0 instead of 3.4.3).

The new sphinx version then fails to build the docs in github, though 
not on the desktop. Exasperation thy name is python environments. 
Holding sphinx at a known working version for now.