ID coordinate attribute added.
This change adds a site_identifier attribute to ID coordinates included
in spot-cubes. This will enable the unique_id_coordinate to be
identified when creating spot-diagnostic cubes without the need to
passing this name in at extraction time. The coordinate attributes will
need to be removed in the spot-extract code so they do not go
downstream; these are intended as internal metadata only.