IMPRO-2026: Additional site identifier for spot-cubes and tweaks to existing metadata (#1458)
* Move spot-data tests up a level to conform with other tests.

* Construction of spot-data cubes with unique ID coordinate.

Adds the functionality to provide an optional unique_site_id list that
will lead to the construction of an additional ID coordinate. This
coordinate can be named or takes a default name of unique_site_id. Unit
tests added for this functionality.

* WMO ID digit padding.

The neighbour finding code has been modified to ensure that the WMO IDs,
when present, are encoded as 5-digit numbers (stored as strings). This
modifies previous behaviour where there was no padding. A unit test has
been updated that captures this behaviour.

* ID coordinate attribute added.

This change adds a site_identifier attribute to ID coordinates included
in spot-cubes. This will enable the unique_id_coordinate to be
identified when creating spot-diagnostic cubes without the need to
passing this name in at extraction time. The coordinate attributes will
need to be removed in the spot-extract code so they do not go
downstream; these are intended as internal metadata only.

* Neighbour finding updated to use unique_site_id.

The neighbour finding code now accepts an optional unique_site_id_name.
This corresponds to an additional ID in the input json site list. Checks
are performed to ensure that the site list is complete and the IDS are
unique.

* Modifications to spot-extract plugin to handle unique site IDs.

This change modifies the spot-extract plugin to include an additional
unique spot ID coordinate on the ouput cube if it is found on the
neighbour cube. Metadata that identifies the unique site ID coordinate
is removed from the final output. Unit and acceptance tests added.

* Slightly extended a test to include a completely missing wmo_id key.

* Review changes.

* Remove redundant 'is None'.

* Update improver_tests/spotdata/test_NeighbourSelection.py

Co-authored-by: Stephen Moseley <stephen.moseley@metoffice.gov.uk>

* Nature of unique ID clarified and enforced.

Added a fuller description of what the unqiue ID values can be to the
CLI and plugin. Added a test to enforce the length of the ID.

Co-authored-by: Stephen Moseley <stephen.moseley@metoffice.gov.uk>