compare_coords function modified to accept ignored_coords list (#1535)
* compare_coords function modified to accept ignored_coords list.

The function now accepts an additional argument that takes the form of a list of coordinate names. This list is checked when cycling through the cubes to ensure that the presence / absence of a coordinate in this list from one or more cubes is not flagged as a difference. Differences in values for an ignored coordinate are also ignored.

* Remove list as argument default.