Modify weather symbols to produce time bounded cubes.
The input cubes are checked for period diagnostics, and these periods
are checked to ensure they all match. If they don't match an exception
is raised. If they do match one of the time-bounded diagnostic cubes is
used as the template for creating the weather symbol cube. This ensures
the time coordinate possesses suitable time bounds.