Spot extraction modified to work with multi-dimensional aux coords.
To be able to spot extract local-timezone maximum temperatures and UV indexes from the global grid the spot-extraction code has been modified to work with 2-D time coordinates. The time coordinate value associated with each spot site is extracted from the 2-D grid and a 1-D coordinate associated with the spot-index is created.