Review changes. Acceptance test added.
The plugin remains relatively slow, and this is dominated by the speed of the mode calculation from scipy. This appears to be the fastest algorithm available, so I cannot do much to speed up this element of the code. There is some x-y looping that could potentially be vectorised to speed up the code in situations when many ties must be resolved, but I have been unable to come up with an acceptable solution to this problem.