MOBT-84: Cell method updates for precipitation phases (#1570)
* Combine multiplication method modifies cell methods appropriately.

The CubeMultiplier plugin has been modified to ensure that cell methods that reference the diagnostic name in the associated comment are updated when new diagnostics are created. This uses the new diagnostic name provided to the plugin as the basis for the updated comment. Those cell methods that do not reference the diagnostic name are left unchanged on the resulting diagnostic cube.

* Unit tests added for the cell method modification within CubeMultiplier.

Some tidying was also done for the existing tests.

* Combine CLI acceptance tests extended to cover cell method modification.

* Review changes.