MOBT-86: Spot extracting from data with 2-dimensional time coordinates (#1573)
* Spot extraction modified to work with multi-dimensional aux coords.

To be able to spot extract local-timezone maximum temperatures and UV indexes from the global grid the spot-extraction code has been modified to work with 2-D time coordinates. The time coordinate value associated with each spot site is extracted from the 2-D grid and a 1-D coordinate associated with the spot-index is created.

* Add units tests for SpotExtraction components.

* Add acceptance test.

* Updated compare CLI acceptance test to work when using new global neighbour cube.

* Review changes.

* Checksum updated.

* Simplified code along lines suggested in review.

* Makes 2-d time coordinates for test with expected units.

This is just to make things consistent between real world usage and 
tests. The existing code to generate time coordinates only deals with 
single dimensions, so here the _create_time_point code is invoked and 
the custom coordinate constructed.