MOBT-77: Weather symbols to represent an extended period (#1552)
* First draft of a modal symbol plugin.

* Grouping solution and rename.

* Adding unit tests.

* Formatting.

* Add type-hinting throughout.

* Review changes. Acceptance test added.

The plugin remains relatively slow, and this is dominated by the speed of the mode calculation from scipy. This appears to be the fastest algorithm available, so I cannot do much to speed up this element of the code. There is some x-y looping that could potentially be vectorised to speed up the code in situations when many ties must be resolved, but I have been unable to come up with an acceptable solution to this problem.

* Doc-string fixes.

Co-authored-by: Stephen Moseley <stephen.moseley@metoffice.gov.uk>

* Extending test coverage.

Acceptance tests extended to cover spot input and to include 
tie-breaking test data for both spot and gridded inputs.
The unit tests have been extended to cover spot data input through an 
additional parameterization.

* Day night working, but needs unit tests.

* Day night fixed to work with partial periods. Unit tests extended to cover this functionality.

* Checksums and formatting.

* Parameterize acceptance tests.

* Checksums updated.

* Remove day-night differentiation in output of daily summary symbols.

* Update doc-strings in plugin and CLI

Co-authored-by: Stephen Moseley <stephen.moseley@metoffice.gov.uk>