Remove cache from scheduled latest environment run
As the action was scheduled to run daily, the conda environment in cache
never expired, so it would never update with newly available package
versions.
This commit removes the cache which will increase time to run these
tasks, but this only affects scheduled tasks rather than tasks triggered
by pushes and pull requests.