MOBT-211: mosg__model_run attribute handling in weather symbols (#1670)
* Factor out set_record_run_attr function.

This will enable the reuse of this functionality, possibly with
modifications, for diagnostics that are created after model blending.

* Update record_run_attr function to avoid duplicate entries.

* Add exception for model blended input without model_run_attr.

It is not possible to create a model_run_attr from previously model
blended inputs if they do not contain a model_run attribute. The cycles
at which these source models were run has been lost. As such the
set_record_run_attr function will raise an exception.

* Add model_run setting for wxcode.

* Add model_run_setting for wxmodal.

* Update checksums to match new test data.

* Additional tests.

Added tests to cover the invocation of the set_record_run_attr function 
from the weather symbols and modal symbol plugins.

* Typo fix.

* Add exception for missing model_id_attr in instances where it is required.

* Add unit tests for new is_model_blended function.

* Review changes.

* Blend test tweak.

* Review changes.

* Update improver/blending/utilities.py

Co-authored-by: gavinevans <gavin.evans@metoffice.gov.uk>

Co-authored-by: gavinevans <gavin.evans@metoffice.gov.uk>