Adds a freezing rain plugin.
This plugin, based upon the snow fraction plugin, calculates the
freezing rain probability. This is done by adding the rain and
sleet probabilities, for rates or accumulations, and multiplying
by the probability that the temperature is below freezing. The
temperature may be a surface temperature, or an air temperature at
the altitude of interest, e.g. screen level. This is not the
formal definition of freezing rain, but it gives an indication of
the likelihood that ice will form.