Adds a freezing rain plugin (#1684)
* Add a probability inversion utility.

This commit adds a utility to invert probabilities and modify the
metadata appropriately; specifically the spp__relative_to_threshold
coordinate attribute and the cube title. This is more than is
required for this work, but I was genuinely surprised we've never
developed such a utility.

* Adds a freezing rain plugin.

This plugin, based upon the snow fraction plugin, calculates the
freezing rain probability. This is done by adding the rain and
sleet probabilities, for rates or accumulations, and multiplying
by the probability that the temperature is below freezing. The
temperature may be a surface temperature, or an air temperature at
the altitude of interest, e.g. screen level. This is not the
formal definition of freezing rain, but it gives an indication of
the likelihood that ice will form.

* Unit tests for FreezingRain plugin.

* Review changes.

* Add realization matching and unit test.

* Changed spp dict to contain namedtuple. Other review changes.

* Review changes.