Correction to the setting of the model config attribute.
Unlike most code where we set the model id attribute, for freezing rain 
in the nowcast period we use a mix of models. To capture all the models 
contributing to the output the method of setting needed to be changed. A 
unit test has been added to catch this case.