MOBT-273: OccurrenceWithinVicinity invoked using a grid point radius. (#1701)
* Modifications to allow OccurrenceWithinVicinity to work using a grid point radius.

This change allows the vicinity plugin to be called directly with a grid 
point radius. Doing so avoids the need to calculate such a radius within 
the code, and thus removes the enforcement that the projection be equal 
areas. This allows us to use a vicinity neighourbourhood in the phase 
change level code for a global lat-lon grid. Doc-strings have been 
updated to indicate that the user must tred carefully here as a grid 
point radius applied to some projections will lead to a very strange 
effective kernel shape.

* Add test for exception when over specifying radius.

* Tweaked checks in the plugin and extended unit test coverage.