Freezing rain calculation to record a single model configuration.
The freezing rain diagnostic produced by the nowcast is constructed using advection precipitation accumulations, but the temperature diagnostic comes from the UKV model. Prior to this change the model_configuration attribute was recording both as sources of data. However it has been decided that the resulting diagnostic is nowcast product and should be recorded as such. This also removes an issue that was manifesting in blending as a result of the original change; the model "ncuk uk_det" is not a single model and no single weight could be ascertained to use when model blending it with other models.