Handle mismatches coordinate dimensions and coordinate system when transforming to lat-lon (#1715)
* Handle mistmatches in units between cube and underlying projection.

The units of the spatial coordinates in an iris cube can have various 
units. The underlying coordinate system is based on proj which has an 
assumed unit for each projection. These are metres for cases that are 
not in degrees, with some possible undocumented exceptions; these are 
likely to be projections we do not encounter commonly. When we can 
embrace a newer version of cartopy, and thus proj, we will be able to 
access the default unit of any given projection. This will enable us to 
do away with the assumption of metres used in this change; TODO added to 
this effect.

* Black fix.