Review changes.
I am no longer able to demonstrate the issue of cell method updates
within the threshold loop leading to cube merging issues. As such
I have reverted the update_metadata method to its original form.

I have modified the behaviour of OccurrenceWithinVicnity to raise an
exception is a radius of 0 is passed in.