MOBT-314: Add radius_of_vicinity coordinate to vicinity processed diagnostics (#1753)
* Remove option to specify no vicinity radius to the occurence within vicinity code.

Slighty refactor the setting of the grid point radius from a distance,
and modify the unit tests to work following this change.

* Metadata setting in occurence within vicinity.

This commit moves the setting of the _in_vicinity name from within the 
threshold CLI to within the OccurenceWithinVicinity plugin. To achieve 
this and add the correct name the input cubes need to be clearly 
probability or not probability cubes. As such the threshold plugin has 
been modified to apply the correct name to each thresholded slice as it 
processes it. Each slice is sent to the vicinity processing 
individually, so the names need to indicate the data type. However, if 
the cell_methods are modified for each slice the resulting cubelist 
cannot be merged. Hence the _update_metadata method in BasicThreshold 
now has options for what to update.

This commit also contains the code to add the radius in vicinity 
coordinate and associated tests. However I've commented out the addition 
of this coordinate and will add it in the next commit. That enables the 
code at this point, with all the changes to Threshold to be run against 
unchanged acceptance test data to demonstrate that it works as expected.

* Add radius in vicinity coordinate.

* Remove vicinity name setting from vicinity CLI.

* Review changes.

I am no longer able to demonstrate the issue of cell method updates
within the threshold loop leading to cube merging issues. As such
I have reverted the update_metadata method to its original form.

I have modified the behaviour of OccurrenceWithinVicnity to raise an
exception is a radius of 0 is passed in.

* Update doc-string

* 2nd review changes. Adds more rigorous testing of radii provided by user.

* Review tweaks