Fix cube name in regridding (#1761)
* Fix cube name in regridding

The regridding code calls create_regrid_cube which generates a new cube
with regridded data. The way in which it does this results in an
"unknown" variable name if the input cube does not have a standard name.
The tests added here fail if used with the original code.

* Set var_name independently

The regridded cube is renamed using the standard_name (if available) or
the long_name (if not). In this change the var_name can be independently
set.

* Add test setting var_name and standard_name

This test sets a different var_name to the standard_name and checks that
both are present.

* Condense tests with pytest parametrise

Co-authored-by: Tom Gale <tom.gale@bom.gov.au>