MOBT-345: Remove expectation of forecast_period coordinate when blending (#1791)
* Remove expectation of forecast_period coordinate when blending.

The weighted blending plugins at the moment presume a forecast_period
coordinate exists, even when it is not being used for blending. This
presumption is not necessary, it largely relates to appending comments
and recalculating the value of the forecast period. This change ensures
that if the forecast_period is not present on the input, and is not
being used in the blending, it is not required.

* Review changes

* Sphinx fix and extended doc-string.