Bug fix for application of emos with validity time filtering.
We have functionality that checks that the forecasts being calibrated
with EMOS are at validity time that we expect to be calibrated. This
is potentially useful when there is training data only available for
specific times / periods that we want to ensure is not inadvertantly
applied to other times / periods.

If a mismatch between the forecast time and the times we deem the
coefficients to be applicable is found, the uncalibrated forecast
is returned. A logic error meant that in cases where a percentile
forecast was provided with a target of probability outputs, in cases
of time mismatches the percentile forecast was returned. This led
to our spot probabilities for tempmax12h and tempmin12h actually
being output as percentiles at all uncalibrated times.

Note that this functionality is a backstop, but it is no longer
strictly required as the suite is only providing coefficients
at the correct times. This means the tolerate_time_mismatch
option would actually be sufficient. We may wish to remove this
in future, or we may wish to simplify the suite and rely on this
more heavily; it is therefore left in place for now.