Treatment of missing values while doing neighbourhood processing
Previously neighbourhood processing did not handle missing data in
the input fields consistently or as expected. It ended up treating
missing data as "0" values rather than missing values. This very
simple change "burns" back in the missing data mask at the very
end of the neighbourhood processing. In other parts of the code
this sort of process is referred to as re-masking.