Implement "proper" blending of wind directions.
We blend wind direction data using the IMPROVER wind direction
functionality. We identify wind direction data by it having units
of degrees.

This identification of wind directions could be generalised to
any variabe which has units that can be converted to degrees, not
just degrees. For example, if the units were radians. But I've not
seen things like radians used in meteorology, so just restricting
the code to variables with units of "degrees" is OK for now.