Wind direction changes (#1958)
* Simplify how angles are converted from (-180, 180] to [0, 360)

The simpler method takes about 65% the CPU time as the original
method (tested on a massive array of angles). Furthermore, the
original method introduces a tiny error for a very few angles near
0/360 degrees.

* Implement "proper" blending of wind directions.

We blend wind direction data using the IMPROVER wind direction
functionality. We identify wind direction data by it having units
of degrees.

This identification of wind directions could be generalised to
any variabe which has units that can be converted to degrees, not
just degrees. For example, if the units were radians. But I've not
seen things like radians used in meteorology, so just restricting
the code to variables with units of "degrees" is OK for now.

* Add proper handling of wind direction to temporal interpolation.

* Fix a temporal interpolation bug.

This was introduced by me at the previous commit.

* Some tests that wind direction interpolation works.

Test 1: 350->20 degrees, expect 5 degrees.
Test 2: 40->60 degrees, expect 50 degrees.

* Put in unit tests to test blending of wind directions

We assume that wind directions can be identified by the units
attribute being "degrees". If this is the case, we use the special
wind direction blending functionality to handle blending.

* Move direction blending into the correct place.

Initially we had direction blending in the wrong function. It
worked, as long as that function was called. If it wasn't then
direction blending didn't happen.

* Change contributor name from Tim to Timothy.

* Remove extra line.

* Fix .mailmap

* Some suggestions from the reviewer.

* For some reason bilinear.py got messed around.

* Covert to float32

* simplify code

* Add keyword.

---------

Co-authored-by: Timothy Hume <timothy.hume@bom.gov.au>
Co-authored-by: Belinda Trotta <btrotta-bom@users.noreply.github.com>