MOBT-1011: Modal symbols bug fix (#2246)
* Remove deprecated method utcfromtimestamp.

* Changes to the modal code handling of model__run_attr.

All input cubes have their model__run_attr modified prior to promotion to a coordinate on the cube and subsequent merging to ensure they match. They are modified to a superset of the entries from all of the input cubes. This ensure all the data is captured but we cannot have alternating coordinate shapes that causes iris to construct oddly shaped cubes that cause us problems.

* Style fixes.

* Update checksums.

* Review changes. Generalise fattribute value extraction function, etc.